+++ out/pages/spriteeditor.js
@@ -119,7 +119,7 @@ class SpriteEditorPage {
                 // Put in a delay to ensure editor is fully loaded before importing project
                 let delay = (!isOpen ? 750 : 5);
                 // Process
-                application.delay(delay).then(_ => this.loadFileContent("loadProject", loadProjectUri));
+                yield application.Delay(delay).then(_ => this.loadFileContent("loadProject", loadProjectUri));
             }
         });
     }